Improve STR repeat counts plot for STRs with high repeat counts
Previously, the STR repeats count plot always drew one bin per repeat
count. For STRs with repeat counts that approached or exceeded the
number of pixels available for the plot (for example, C9ORF72), this
resulted in the x axis appearing as a solid rectangle.

Fewer pixels / bin also made it more difficult to hover over a bin to
see the number of alleles in that bin.