Catch variants in ClinvarAllVariantsPlot that run past 3' UTR
In some cases, such as one variant on BRCA1, the data from VEP gives us an HGVSP for a variant that runs past the end of the 3' UTR as specified by Gencode. These were causing the plot of Clinvar variants to crash.

Here we fix that, along with an assortment of off-by-one errors and edge cases.