Better align StructuralVariant type with GraphQL API
The `StructuralVariant` type in the frontend code actually corresponds to `STructuralVariantDetails` on the API side, rather than the type of the same name. Here we update the frontend type to match the schema for the API type. Doing so made at least one bug visible to the type checker (the possibility of a null `evidence` field), and allowed a lot more general cleanup of this part of the code.