Clean up CopyNumberVariant associated types
This is prep to changing its "populations" field to "ancestry_groups".

Note that, following the GraphQL API schema, "length" and other fields that don't seem like they should be null-able can be null. We should check the actual data, ensure there's no nulls in the fields in question, and assuming not tighten these types up.