Split exome and genome frequencies in VA data
VA data was previously always using exome frequencies if available, genome if not. However, given that this is a general-purpose API, it's hard to say if users will want exomes, genomes, or both for their use cases. Hence, rather than try to make that decision for them, we allow (and require) them now to say which they want in their queries.