Comment on places where we can use useTableSort in the future
The logic copypasta'ed into these three tables is an older version of that which was just extracted into useTableSort. It's not possible to use hooks in a class component like these three, but we leave a note in these places reminding us to use useTableSort in case we convert them to functional components in the future.